API: Some refactor for endpoints and base application.
