+++ api/exdemon/constants.py
@@ -1,0 +1,16 @@
+
+# Global constants
+
+# HTTP Codes
+OK = 200
+CREATED = 201
+ACCEPTED = 202
+NO_CONTENT = 204
+BAD_REQUEST = 400
+UNAUTHORIZED = 401
+FORBIDDEN = 403
+NOT_FOUND = 404
+NOT_ALLOWED = 405
+CONFLICT = 409
+INTERNAL_ERROR = 500
+NOT_IMPLEMENTED = 501
